add some shit

+++ src/general_support/integrator/Integrator.java
@@ -8,7 +8,7 @@ public interface Integrator {
      *  @param position Vector indicating current position
      * @param velocity Vector indicating current velocity
      * @param acceleration Vector indicating current acceleration
-     * @param timeStep
+     * @param timeStep timestep
      */
-    void integrate(Vector position, Vector velocity, Vector acceleration, int timeStep);
+    void integrate(Vector position, Vector velocity, Vector acceleration, double timeStep);
 }
